Welcome to CraftIRC5!
Basically a whole rewrite, with 55 changed files, it's a LOT of changes.
And the changes aren't small, outside from a drastically changed code structure,
it provides a different config file and better readabillity for me, which
would likely matter the most since I am sure that I will be the only one to work
on the project. However contributions from others are welcome.

Now, CraftIRC5 is only something that is done by me to just prove the significance
of the changes, it isn't something that is happening outside the repo, unless
mbaxter believes that these changes were wonderfull (I doubt that).
It might come with better flexibillity (although, who know's?) but it certainly
removed a lot of features as of now.

Plus, most of these things haven't been tested by me nor by anyone else, however
they will be in the coming days, so don't worry too much and help with the debugging.

Signed-off-by: Geolykt <admin@geolykt.de>