Started porting to forge
It will be a momumental task, but since I already got some experience with porting this, it shouldn't be all too hard.
However for today (well, at least for the next few hours) I will stop working on it.
The main issue will be configurations, but are they important? I got no idea.

Signed-off-by: Geolykt <admin@geolykt.de>